+++ Sources/SwiftUIKit/Pickers/MultiPicker.swift
@@ -1,0 +1,20 @@
+//
+//  SwiftUIView.swift
+//  SwiftUIView
+//
+//  Created by Daniel Saidi on 2021-08-21.
+//
+
+import SwiftUI
+
+struct SwiftUIView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        SwiftUIView()
+    }
+}